fix warnings identified by clang -Weverything

See merge request ewoo/youtube-unthrottle!78

+++ main.c
@@ -117,7 +117,7 @@ err:
 		return false;
 	}
 
-	to_stderr("%s \"%s\" at offset %zd: %s", action, str, (size_t)loc, err);
+	to_stderr("%s \"%s\" at offset %zu: %s", action, str, loc, err);
 	return false;
 }
